Update README.md
Run this pair of programs on a plain text file with the following line of code:
$ cat ../textfile | Python3 ./mapper_hmwk20.py | sort | Python3 ./reducer_hmwk20.py

Being able to explicitly define what happens to each part of the text would be extremely useful in natural language processing. Coding shows you what is happening "under the hood" in python libraries like NLTK and SpaCy, and can give you  the flexibility to add special treatment for particular irregularities in a particular text.